% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-build-base.R
\name{build_matrix}
\alias{build_matrix}
\title{Build a SpatRaster from vapour vector}
\usage{
build_matrix(p, v)
}
\arguments{
\item{p}{ezgrid object}

\item{v}{list of numeric vectors from vapour}
}
\value{
matrix
}
\description{
Build a SpatRaster from vapour vector
}
