% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-build-SpatRaster.R
\name{build_SpatRaster}
\alias{build_SpatRaster}
\title{Build a SpatRaster from vapour vector}
\usage{
build_SpatRaster(p, v)
}
\arguments{
\item{p}{params from ezwarp}

\item{v}{a numeric vector with raster values}
}
\value{
a SpatRaster
}
\description{
Build a SpatRaster from vapour vector
}
