% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buid_stars.R
\name{build_stars}
\alias{build_stars}
\title{Build a stars object from vapour vector}
\usage{
build_stars(p, v)
}
\arguments{
\item{p}{params from ezwarp}

\item{v}{a numeric vector with raster values}
}
\value{
a stars object
}
\description{
Build a stars object from vapour vector
}
