% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-build-base.R
\name{build_vector}
\alias{build_vector}
\title{Build a list of vectors from format}
\usage{
build_vector(p, v)
}
\arguments{
\item{p}{ezgrid object}

\item{v}{list of numeric vectors from vapour}
}
\value{
matrix
}
\description{
Build a list of vectors from format
}
