% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-warp.R
\name{build_warp_inputs}
\alias{build_warp_inputs}
\title{list raster sources}
\usage{
build_warp_inputs(x, y, res)
}
\arguments{
\item{x}{}

\item{y}{}

\item{res}{}
}
\value{

}
\description{
list raster sources
}
