% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_grid_form}
\alias{check_grid_form}
\title{check input template source format/length}
\usage{
check_grid_form(x)
}
\arguments{
\item{x}{}
}
\value{
list
}
\description{
ensures template is of length 1
}
