% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-inputs.R
\name{check_res_form}
\alias{check_res_form}
\title{check that res is provided if template is a spatial vector}
\usage{
check_res_form(x, .res)
}
\arguments{
\item{x}{}
}
\value{
list
}
\description{
ensures res is provided where spatial vector provided
}
