% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_source.R
\name{check_source}
\alias{check_source}
\alias{check_source.SpatRaster}
\alias{check_source.stars}
\alias{check_source.stars_proxy}
\alias{check_source.character}
\alias{check_source.sf}
\alias{check_source.sfc}
\alias{check_source.SpatVector}
\title{Check the source raster classes}
\usage{
check_source(s)

\method{check_source}{SpatRaster}(s)

\method{check_source}{stars}(s)

\method{check_source}{stars_proxy}(s)

\method{check_source}{character}(s)

\method{check_source}{sf}(s)

\method{check_source}{sfc}(s)

\method{check_source}{SpatVector}(s)
}
\arguments{
\item{s}{a raster source - either character i.e a local source or remote.
If remote, src must be prefixed with relevant gdal module. e.g. "/vsicurl/"}
}
\value{
character source.
}
\description{
If a SpatRaster is provided the source of the raster is used. if it is in
memory then and error is thrown. if character provided then assumed to be
correct.
}
