% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-warp.R
\name{dims_from_box}
\alias{dims_from_box}
\title{get the target x y dimensions from bbox and desired res.}
\usage{
dims_from_box(.box, .res)
}
\arguments{
\item{.box}{}

\item{.res}{}
}
\value{
numeric vector xy dims
}
\description{
get the target x y dimensions from bbox and desired res.
}
