% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezwarp-package.R
\docType{package}
\name{ezwarp-package}
\alias{ezwarp-package}
\title{ezwarp: A utility tool for working with gdal warp in R}
\description{
The goal of ezwarp is to simplify the use of gdalwarp in R. gdalwarp is a very powerful and fast Image reprojection and warping utility. It can be used to reproject (i.e. transform to a new Coordinate Reference System), crop, mask and resample. Unlike many tools in R, this can all be done in a single command, using the warp tool, saving time and code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Permian-Global-Research/ezwarp}
  \item \url{https://permian-global-research.github.io/ezwarp/}
  \item Report bugs at \url{https://github.com/Permian-Global-Research/ezwarp/issues}
}

}
\author{
\strong{Maintainer}: Hugh Graham \email{hugh.graham@permianglobal.com} (\href{https://orcid.org/0000-0001-9451-5010}{ORCID})

}
\keyword{internal}
