% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_dim.R
\name{get_dim}
\alias{get_dim}
\alias{get_dim.SpatRaster}
\alias{get_dim.stars}
\alias{get_dim.stars_proxy}
\alias{get_dim.chracter}
\alias{get_dim.ezgrid}
\title{get_proj: get projection of a spatial object}
\usage{
get_dim(x, ...)

\method{get_dim}{SpatRaster}(x, ...)

\method{get_dim}{stars}(x, ...)

\method{get_dim}{stars_proxy}(x, ...)

\method{get_dim}{chracter}(x, ...)

\method{get_dim}{ezgrid}(x, ...)
}
\arguments{
\item{x}{...}
}
\description{
get projection of spatial object
}
\details{
get projection of spatial object
}
