% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-proj.R
\name{get_proj}
\alias{get_proj}
\alias{get_proj.SpatRaster}
\alias{get_proj.SpatVector}
\alias{get_proj.sf}
\alias{get_proj.sfc}
\alias{get_proj.stars}
\alias{get_proj.stars_proxy}
\alias{get_proj.ezgrid}
\alias{get_proj.character}
\title{Get Spatial Projection}
\usage{
get_proj(x, ...)

\method{get_proj}{SpatRaster}(x, ...)

\method{get_proj}{SpatVector}(x, ...)

\method{get_proj}{sf}(x, ...)

\method{get_proj}{sfc}(x, ...)

\method{get_proj}{stars}(x, ...)

\method{get_proj}{stars_proxy}(x, ...)

\method{get_proj}{ezgrid}(x, ...)

\method{get_proj}{character}(x, ...)
}
\arguments{
\item{x}{A spatial object, file path or source}

\item{...}{Not used}
}
\value{
A character - WKT projection string.
}
\description{
A class agnostic function to return the projection of a spatial object or
source. returned projection uses wkt format.
}
\examples{
f <- system.file("ex/elev.tif", package = "terra")
get_proj(f)
get_proj(terra::rast(f))
f2 <- system.file("ex/lux.shp", package = "terra")
get_proj(f2)

}
\seealso{
Other spatial helpers (class agnostic): 
\code{\link{get_dim}()},
\code{\link{get_ext}()},
\code{\link{get_source}()}
}
\concept{spatial helpers (class agnostic)}
