% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-build-base.R
\name{matrix_thing}
\alias{matrix_thing}
\title{Internal function which is used to form a matrix from single vector band}
\usage{
matrix_thing(.v, .p)
}
\arguments{
\item{.v}{vector}

\item{.p}{ezgrid object}
}
\value{
Rotated matrix
}
\description{
Internal function which is used to form a matrix from single vector band
}
\keyword{internal}
