% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read_spatial.R
\name{read_spat_info}
\alias{read_spat_info}
\title{read a spatial url/path}
\usage{
read_spat_info(x)
}
\arguments{
\item{x}{A character vector - source for a raster or spatial vector}
}
\value{

}
\description{
read a spatial url/path
}
