% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-warp.R
\name{round_bbox}
\alias{round_bbox}
\title{round a bbox to desired interval}
\usage{
round_bbox(.box, .res)
}
\arguments{
\item{.box}{}

\item{.res}{}
}
\value{
vector with bbox dims
}
\description{
round a bbox to desired interval
}
