% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup-warp.R
\name{round_nearest}
\alias{round_nearest}
\title{Round to nearest value}
\usage{
round_nearest(x, accuracy, f = round)
}
\arguments{
\item{x}{numeric}

\item{accuracy}{numeric. Target multiple to round to.}

\item{f}{}
}
\value{

}
\description{
A literal copy of \code{plyr::round_any()}
}
