% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciplot.R
\name{sciplot}
\alias{sciplot}
\alias{sciplot.SpatRaster}
\alias{sciplot.stars}
\alias{sciplot_pals}
\title{sciplot: plot methods using colours from \code{{scico}}}
\usage{
sciplot(x, pal = "acton", n, direction, centre, ...)

\method{sciplot}{SpatRaster}(
  x,
  pal = "acton",
  n = 256,
  direction = 1,
  centre = FALSE,
  n_quantile = NULL,
  ...
)

\method{sciplot}{stars}(x, pal = "acton", n = 11, direction = 1, centre = FALSE, ...)

sciplot_pals()
}
\arguments{
\item{x}{object to plot, either a SpatRaster or stars object}

\item{pal}{Default is 'acton'. Character describing colour palette to use.
choose from \code{sciplot_pals()}}

\item{n}{Numeric of length 1. The number of breaks/colours to include in the
sclae}

\item{direction}{Default is 1. Direction of colour palette to reverse use -1.}

\item{centre}{Logical. If TRUE, the palette is centred around zero. good for
bivariate/split palettes.}

\item{...}{additional args passed to plot.}

\item{n_quantile}{Numeric. if not NULL sets the number of quantile breaks to
use for sciplot.SpatRaster. See details.}
}
\value{
character vector of scico palette names.
see \code{scico::scico_palette_names()}
}
\description{
plot a 'SpatRaster' or 'stars' object using \code{{scico}} palettes. View palette
options with \code{sciplot_pals()}
}
\details{
n_quantile sets quantile breaks for the colour palette. When using breaks,
terra::plot does not support continuous legend as is the default in stars.

sciplot_pals plots the various sciplot palettes and returns their names as a
character vector. Wrapper for \code{scico::scico_palette_show()}
and \code{scico::scico_palette_names()}.
}
\examples{
f <- system.file("ex/elev.tif", package = "terra")
r.stars <- stars::read_stars(f)
sciplot(r.stars)

sciplot_pals()
}
\concept{plot_helpers}
