% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp-engines.R
\name{sf_warp_util}
\alias{sf_warp_util}
\title{Easy Warp using sf::gdalutils}
\usage{
sf_warp_util(params, destination, resample, compression, options, ...)
}
\arguments{
\item{params}{A list containing the raster source(s), ext, dim and crs.}

\item{destination}{out put destination. if NUll then a tempfile is created}

\item{resample}{resampling method}

\item{compression}{the tif compression method to use - e.g. "DEFLATE" or "LZW"}

\item{options}{character vector with gdal options.}

\item{...}{Not used.}
}
\value{

}
\description{
stable and supports >memory rasters and cutline feature.
}
