% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfwarputil.R
\name{sfwarp_util}
\alias{sfwarp_util}
\title{Easy Warp using sf::gdalutils}
\usage{
sfwarp_util(
  x,
  y,
  res,
  bands = 1L,
  resample = "bilinear",
  cutline = NULL,
  destination = NULL,
  dstnodata = -999,
  ...
)
}
\arguments{
\item{x}{a raster source or SpatRaster}

\item{y}{a raster source, SpatRaster, sf, or sfc}

\item{res}{numeric. the resolution of the output SpatRaster.}

\item{bands}{integer. which bands to use from the source}

\item{resample}{resampling method}

\item{cutline}{an sf or ogr-readable spatial vector source to mask the output raster.}

\item{destination}{out put destination. if NUll then a tempfile is created}

\item{...}{Not used.}
}
\value{

}
\description{
stable and supports >memory rasters and cutline feature.
}
