% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp-engines.R
\name{vapour_warp_util}
\alias{vapour_warp_util}
\title{vapour_warp_util}
\usage{
vapour_warp_util(params, bands, resample, opts, ...)
}
\arguments{
\item{params}{A list containing the raster source(s), ext, dim and crs.}

\item{bands}{numeric veor of bands to include.}

\item{resample}{resampling method}

\item{opts}{character vector with gdal options.}

\item{...}{passed to \code{vapour::vapour_warp_raster}}
}
\value{

}
\description{
vapour_warp_util
}
