% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezwarp-package.R
\docType{package}
\name{ezwarp-package}
\alias{ezwarp}
\alias{ezwarp-package}
\title{ezwarp: A utility tool for working with gdal warp in R}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://permian-global-research.github.io/ezwarp/}
}

}
\author{
\strong{Maintainer}: Hugh Graham \email{hugh.graham@permianglobal.com} (\href{https://orcid.org/0000-0001-9451-5010}{ORCID})

}
\keyword{internal}
