% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezwarp.R
\name{ezwarp}
\alias{ezwarp}
\title{An easy way to use gdal warp in R}
\usage{
ezwarp(
  x,
  y,
  res,
  bands = NULL,
  resample = "bilinear",
  cutline = NULL,
  crop_to_cutline = FALSE,
  nodata = NULL,
  out_class = c("SpatRaster", "stars", "matrix", "vector"),
  filename = NULL,
  overwrite = TRUE,
  options = "",
  compression = "DEFLATE",
  engine = c("vapour", "sf"),
  ...
)
}
\arguments{
\item{x}{a list or vector of raster source(s) or SpatRaster(s)}

\item{y}{a raster source, SpatRaster, sf, or sfc}

\item{res}{numeric. the resolution of the output SpatRaster.}

\item{bands}{numeric which bands to use from the source. Only used if \code{engine="vapour"}}

\item{resample}{resampling method. default is bilinear, see details.}

\item{cutline}{an sf, sfc, SpatVector or ogr-readable spatial vector source to mask the output raster. see -cutline argument in gdalwarp}

\item{crop_to_cutline}{logical. If TRUE, then the output will be cropped to the limits of the mask given in cutline.}

\item{nodata}{Numeric. No data value to be used for output.}

\item{out_class}{default is "SpatRaster". Can be either "SpatRaster", "stars", "matrix", "vector"}

\item{filename}{the filepath for the out raster. if given and "vapour" is used
for the engine, then the output SpatRaster/stars object will have a source.
If NULL then an in memory raster is returned. If the sf engine is used and
filename is NULL then a tempfile is used.}

\item{overwrite}{logical - should a file be overwritten.}

\item{options}{gdal options.}

\item{compression}{default is "DEFLATE". character describing tif compression e.g. "LZW"}

\item{engine}{either "vapour" or "sf". choose which warper to use. Only vapour
supports in memory raster creation.}

\item{...}{Additional args passed to \code{vapour::vapour_warp_raster}. Might be removed.}
}
\value{

}
\description{
super handy warper but limited in that returns in memory SpatRast. May cause
issues when creating a raster > memory. Also not clear how to implement
the cutline feature.
}
\details{
When selecting the resample method, choose from the following options:
'bilinear' (the default), 'near', 'cubic', 'cubicspline', 'lanczos', 'average', 'mode',
'max', 'min', 'med', 'q1', 'q3', 'sum'. For details on these methods see: https://gdal.org/programs/gdalwarp.html
If raster source is categorical make sure to use 'nearest'.

If "matrix" is used for \code{out_class}, then a matrix is returned if only one band
is targeted. If multiple bands are targeted, then a n-dimensional array is returned.
The matrix is returned oriented North is up.

If "vector" is used for \code{out_class}, a vector is returned for a single band target,
and a list of vectors is returned for a multiband target.
}
\examples{
f <- system.file("ex/elev.tif", package="terra") 
r.terra <- terra::rast(f)
r1a <- ezwarp(f,f, res=1e-4)
r1b <- ezwarp(r.terra,f, res=1e-4, engine = 'sf')
}
