% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezwarp.R
\name{ezwarp}
\alias{ezwarp}
\title{Easy Warper using vapour}
\usage{
ezwarp(x, y, res, bands = NULL, resample = "bilinear", cutline = FALSE, ...)
}
\arguments{
\item{x}{a list or vector of raster source(s) or SpatRaster(s)}

\item{y}{a raster source, SpatRaster, sf, or sfc}

\item{res}{numeric. the resolution of the output SpatRaster.}

\item{bands}{numeric which bands to use from the source.}

\item{resample}{resampling method}

\item{...}{Additional args passed to \code{vapour::vapour_warp_raster}}

\item{mask}{an sf or ogr-readable spatial vector source to mask the output raster. see -cutline argument in gdalwarp}
}
\value{

}
\description{
super handy warper but limited in that returns in memory SpatRast. May cause
issues when creating a raster > memory. Also not clear how to implement
the cutline feature.
}
