% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-dim.R
\name{get_dim}
\alias{get_dim}
\alias{get_dim.SpatRaster}
\alias{get_dim.SpatVector}
\alias{get_dim.sf}
\alias{get_dim.sfc}
\alias{get_dim.stars}
\alias{get_dim.stars_proxy}
\alias{get_dim.character}
\alias{get_dim.ezgrid}
\title{Get spatial Dimension}
\usage{
get_dim(x, ...)

\method{get_dim}{SpatRaster}(x, ...)

\method{get_dim}{SpatVector}(x, ...)

\method{get_dim}{sf}(x, ...)

\method{get_dim}{sfc}(x, ...)

\method{get_dim}{stars}(x, ...)

\method{get_dim}{stars_proxy}(x, ...)

\method{get_dim}{character}(x, ...)

\method{get_dim}{ezgrid}(x, ...)
}
\arguments{
\item{x}{A spatial object, file path or source}

\item{...}{Not used}
}
\description{
A class agnostic function to return the dimensions of a spatial object. Only
applicable for rasters - spatial vectors will return NULL.
}
\examples{
f <- system.file("ex/elev.tif", package="terra") 
get_dim(f)
get_dim(terra::rast(f))
f2 <- system.file("ex/lux.shp", package="terra") 
get_dim(f2)

}
\seealso{
Other spatial helpers (class agnostic): 
\code{\link{get_ext}()},
\code{\link{get_proj}()},
\code{\link{get_source}()}
}
\concept{spatial helpers (class agnostic)}
