% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-ext.R
\name{get_ext}
\alias{get_ext}
\title{get_ext: get projection of a spatial object}
\usage{
get_ext(x, ...)
}
\arguments{
\item{x}{...}
}
\description{
get extent of spatial object
}
\details{
get extent of spatial object
}
