% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-ext.R, R/utils-get-proj.R
\name{get_ext.SpatRaster}
\alias{get_ext.SpatRaster}
\alias{get_ext.SpatVector}
\alias{get_ext.sf}
\alias{get_ext.sfc}
\alias{get_ext.stars}
\alias{get_ext.stars_proxy}
\alias{get_ext.ezgrid}
\alias{get_ext.character}
\alias{get_proj}
\alias{get_proj.SpatRaster}
\alias{get_proj.SpatVector}
\alias{get_proj.sf}
\alias{get_proj.sfc}
\alias{get_proj.stars}
\alias{get_proj.stars_proxy}
\alias{get_proj.ezgrid}
\alias{get_proj.character}
\title{get_proj: get projection of a spatial object}
\usage{
\method{get_ext}{SpatRaster}(x, ...)

\method{get_ext}{SpatVector}(x, ...)

\method{get_ext}{sf}(x, ...)

\method{get_ext}{sfc}(x, ...)

\method{get_ext}{stars}(x, ...)

\method{get_ext}{stars_proxy}(x, ...)

\method{get_ext}{ezgrid}(x, ...)

\method{get_ext}{character}(x, ...)

get_proj(x, ...)

\method{get_proj}{SpatRaster}(x, ...)

\method{get_proj}{SpatVector}(x, ...)

\method{get_proj}{sf}(x, ...)

\method{get_proj}{sfc}(x, ...)

\method{get_proj}{stars}(x, ...)

\method{get_proj}{stars_proxy}(x, ...)

\method{get_proj}{ezgrid}(x, ...)

\method{get_proj}{character}(x, ...)
}
\arguments{
\item{x}{...}
}
\description{
get projection of spatial object
}
\details{
get projection of spatial object
}
