% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-source.R
\name{get_source}
\alias{get_source}
\alias{get_source.SpatRaster}
\alias{get_source.stars}
\alias{get_source.stars_proxy}
\alias{get_source.character}
\alias{get_source.sf}
\alias{get_source.sfc}
\alias{get_source.SpatVector}
\title{Check the source raster classes}
\usage{
get_source(s, force = FALSE)

\method{get_source}{SpatRaster}(s, force = FALSE)

\method{get_source}{stars}(s)

\method{get_source}{stars_proxy}(s, force = FALSE)

\method{get_source}{character}(s)

\method{get_source}{sf}(s)

\method{get_source}{sfc}(s)

\method{get_source}{SpatVector}(s)
}
\arguments{
\item{s}{a raster source - either character i.e a local source or remote.
If remote, src must be prefixed with relevant gdal module. e.g. "/vsicurl/"}

\item{force}{Logical to determine if the SpatRaster or stars_proxy object should
be resaved even if on disk sources exist.}
}
\value{
character source.
}
\description{
If a SpatRaster is provided the source of the raster is used. if it is in
memory then and error is thrown. if character provided then assumed to be
correct.
}
