% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciplot.R
\name{sciplot}
\alias{sciplot}
\alias{sciplot.SpatRaster}
\title{sciplot: generic plot but using palettes from {scico}}
\usage{
sciplot(x, ...)

\method{sciplot}{SpatRaster}(x, .pal = "bilbao", .n = 256, direction = 1, ...)
}
\arguments{
\item{x}{...}
}
\description{
plot using {scico} palettes
}
\details{
plot using {scico} palettes
}
