% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp-engines.R
\name{sf_warp_util}
\alias{sf_warp_util}
\title{Easy Warp using sf::gdalutils}
\usage{
sf_warp_util(params, destination, resample, compression, options, ...)
}
\arguments{
\item{destination}{out put destination. if NUll then a tempfile is created}

\item{resample}{resampling method}

\item{compression}{the tif compression method to use - e.g. "DEFLATE" or "LZW"}

\item{...}{Not used.}

\item{x}{a raster source or SpatRaster}

\item{y}{a raster source, SpatRaster, sf, or sfc}

\item{res}{numeric. the resolution of the output SpatRaster.}

\item{bands}{integer. which bands to use from the source}

\item{cutline}{an sf or ogr-readable spatial vector source to mask the output raster.}
}
\value{

}
\description{
stable and supports >memory rasters and cutline feature.
}
