% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp-engines.R
\name{vapour_warp_util}
\alias{vapour_warp_util}
\title{vapour_warp_util}
\usage{
vapour_warp_util(params, bands, resample, opts, ...)
}
\arguments{
\item{params}{}

\item{bands}{}

\item{resample}{}

\item{opts}{}

\item{...}{}
}
\value{

}
\description{
vapour_warp_util
}
